data-integration/push-article-to-firestore/requirements.txt
grpcio
google-cloud-core
google-cloud-firestore